# api/openapi.yaml
openapi: 3.0.3
info:
  title: LecodeKanban API
  version: 0.3.0
  description: |
    REST API of LecodeKanban. Source of truth for Go DTOs and the TypeScript client (`make gen`).

    **Auth:** httpOnly cookies — `lk_at` (access, ~15 min) and `lk_rt` (rotating refresh,
    path-scoped to `/api/v1/auth`). **CSRF:** every non-GET request must echo the `lk_csrf`
    cookie in the `X-CSRF-Token` header (obtain it via `GET /auth/csrf`).

    **Errors** are `{"error": {"code", "message", "fields"?, "meta"?}}`. `code` is stable and
    translated by clients; `message` is an English developer hint. See `GET /i18n/error-codes`.
servers:
  - url: /api/v1
...
paths:
...
  /workspaces/{workspaceId}/cards:
    parameters: [{ $ref: "#/components/parameters/WorkspaceId" }]
    get:
      tags: [cards]
      operationId: listCards
      parameters:
        - $ref: "#/components/parameters/CardStatus"
        - $ref: "#/components/parameters/CardProjectId"
        - $ref: "#/components/parameters/CardAssigneeId"
        - $ref: "#/components/parameters/CardPriority"
        - $ref: "#/components/parameters/CardLabelId"
        - $ref: "#/components/parameters/CardQuery"
        - $ref: "#/components/parameters/CardDue"
        - { name: sort, in: query, schema: { type: string, enum: [key, title, assignee, project, progress, deadline, priority, position, updated] } }
        - { name: order, in: query, schema: { type: string, enum: [asc, desc] } }
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: Cards page
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CardPage" }
    post:
      tags: [cards]
      operationId: createCard
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CardInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Card" }
        "422": { $ref: "#/components/responses/Error" }
  /workspaces/{workspaceId}/cards/summary:
    parameters: [{ $ref: "#/components/parameters/WorkspaceId" }]
    get:
      tags: [cards]
      operationId: cardStatusCounts
      parameters:
        - $ref: "#/components/parameters/CardProjectId"
        - $ref: "#/components/parameters/CardAssigneeId"
        - $ref: "#/components/parameters/CardPriority"
        - $ref: "#/components/parameters/CardLabelId"
        - $ref: "#/components/parameters/CardQuery"
        - $ref: "#/components/parameters/CardDue"
      responses:
        "200":
          description: Card counts per status for the same filters as listCards
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StatusCounts" }
  /workspaces/{workspaceId}/cards/bulk:
    parameters: [{ $ref: "#/components/parameters/WorkspaceId" }]
    post:
      tags: [cards]
      operationId: bulkCards
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/BulkCardAction" }
      responses:
        "200":
          description: Number of cards changed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BulkResult" }
  /workspaces/{workspaceId}/cards/stats:
    parameters: [{ $ref: "#/components/parameters/WorkspaceId" }]
    get:
      tags: [cards]
      operationId: cardStats
      parameters:
        - { name: days, in: query, schema: { type: integer, minimum: 7, maximum: 31 } }
      responses:
        "200":
          description: Dashboard KPIs, daily activity and recent transitions
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DashboardStats" }
  /cards/{cardId}:
    parameters: [{ $ref: "#/components/parameters/CardId" }]
    get:
      tags: [cards]
      operationId: getCard
      responses:
        "200":
          description: Card
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Card" }
        "404": { $ref: "#/components/responses/Error" }
    patch:
      tags: [cards]
      operationId: updateCard
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CardPatch" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Card" }
        "409": { $ref: "#/components/responses/Error" }
    delete:
      tags: [cards]
      operationId: deleteCard
      responses:
        "204": { description: Archived }
  /cards/{cardId}/move:
    parameters: [{ $ref: "#/components/parameters/CardId" }]
    post:
      tags: [cards]
      operationId: moveCard
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CardMove" }
      responses:
        "200":
          description: Moved
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Card" }
        "409": { $ref: "#/components/responses/Error" }
...
  /workspaces/{workspaceId}/cards/board:
    parameters: [{ $ref: "#/components/parameters/WorkspaceId" }]
    get:
      tags: [cards]
      operationId: boardCards
      description: Every card matching the filters in board order (position), capped at 3000.
      parameters:
        - $ref: "#/components/parameters/CardProjectId"
        - $ref: "#/components/parameters/CardAssigneeId"
        - $ref: "#/components/parameters/CardPriority"
        - $ref: "#/components/parameters/CardLabelId"
        - $ref: "#/components/parameters/CardQuery"
        - $ref: "#/components/parameters/CardDue"
      responses:
        "200":
          description: Board cards
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CardBoard" }
  /workspaces/{workspaceId}/labels:
    parameters: [{ $ref: "#/components/parameters/WorkspaceId" }]
    get:
      tags: [cards]
      operationId: listLabels
      responses:
        "200":
          description: Workspace labels
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/Label" } }
    post:
      tags: [cards]
      operationId: createLabel
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LabelInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Label" }
        "409": { $ref: "#/components/responses/Error" }
  /labels/{labelId}:
    parameters: [{ $ref: "#/components/parameters/LabelId" }]
    patch:
      tags: [cards]
      operationId: updateLabel
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LabelPatch" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Label" }
    delete:
      tags: [cards]
      operationId: deleteLabel
      responses:
        "204": { description: Deleted (removed from every card) }
  /cards/{cardId}/checklist:
    parameters: [{ $ref: "#/components/parameters/CardId" }]
    get:
      tags: [cards]
      operationId: listChecklist
      responses:
        "200":
          description: Items in order
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/ChecklistItem" } }
    post:
      tags: [cards]
      operationId: addChecklistItem
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ChecklistItemInput" }
      responses:
        "201":
          description: Created (card progress is recomputed)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ChecklistItem" }
  /checklist-items/{itemId}:
    parameters: [{ $ref: "#/components/parameters/ItemId" }]
    patch:
      tags: [cards]
      operationId: updateChecklistItem
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ChecklistItemPatch" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ChecklistItem" }
    delete:
      tags: [cards]
      operationId: deleteChecklistItem
      responses:
        "204": { description: Deleted }
...
components:
  securitySchemes:
    cookieAuth: { type: apiKey, in: cookie, name: lk_at }
  parameters:
...
    WorkspaceId:
      name: workspaceId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    CardId:
      name: cardId
      in: path
      required: true
      schema: { type: string, format: uuid }
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1 }
    PageSize:
      name: pageSize
      in: query
      schema: { type: integer, minimum: 1, maximum: 500 }
    CardStatus: { name: status, in: query, schema: { $ref: "#/components/schemas/TaskStatus" } }
    CardProjectId: { name: projectId, in: query, schema: { type: string, format: uuid } }
    CardAssigneeId: { name: assigneeId, in: query, schema: { type: string, format: uuid } }
    CardPriority: { name: priority, in: query, schema: { $ref: "#/components/schemas/Priority" } }
    CardQuery: { name: q, in: query, schema: { type: string }, description: Search in title or key (e.g. LK-12) }
    CardLabelId: { name: labelId, in: query, schema: { type: string, format: uuid } }
...
    LabelId: { name: labelId, in: path, required: true, schema: { type: string, format: uuid } }
    ItemId: { name: itemId, in: path, required: true, schema: { type: string, format: uuid } }
...
    CardDue: { name: due, in: query, schema: { type: string, enum: [overdue, today, week, month, none] } }
  responses:
...
    Error:
      description: Error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string, example: auth.invalid_credentials }
            message: { type: string }
            fields:
              type: array
              items: { $ref: "#/components/schemas/FieldError" }
            meta:
              type: object
              additionalProperties: true
    FieldError:
      type: object
      required: [field, code]
      properties:
        field: { type: string }
        code: { type: string, example: min_length }
        params: { type: object, additionalProperties: true }
...
    TaskStatus:
      type: string
      enum: [todo, in_progress, in_review, done]
    Priority:
      type: string
      enum: [high, medium, low]
...
    Tone:
      type: string
      enum: [teal, amber, purple, red, neutral]
    ProjectIcon:
      type: string
      enum: [folder, rocket, shield, code, target, sparkles, globe, megaphone, layers, flask]
    PersonRef:
      type: object
      required: [id, name, avatarUrl]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        avatarUrl: { type: string, nullable: true }
    ProjectRef:
      type: object
      required: [id, key, name, tone, icon]
      properties:
        id: { type: string, format: uuid }
        key: { type: string }
        name: { type: string }
        tone: { $ref: "#/components/schemas/Tone" }
        icon: { $ref: "#/components/schemas/ProjectIcon" }
...
    Card:
      type: object
      required: [id, key, number, title, description, status, columnId, priority, progress, dueDate, project, assignees, labels, checklist, commentCount, attachmentCount, position, version, createdAt, updatedAt, completedAt]
      properties:
        id: { type: string, format: uuid }
        key: { type: string, example: LK-12 }
        number: { type: integer }
        title: { type: string }
        description: { type: string }
        status: { $ref: "#/components/schemas/TaskStatus" }
        columnId: { type: string, format: uuid }
        priority: { $ref: "#/components/schemas/Priority" }
        progress: { type: integer, minimum: 0, maximum: 100, description: "Derived: checklist share, else workflow stage; 100 only when done (ADR 0010)" }
        dueDate: { type: string, format: date, nullable: true }
        project: { $ref: "#/components/schemas/ProjectRef" }
        assignees: { type: array, items: { $ref: "#/components/schemas/PersonRef" } }
        labels: { type: array, items: { $ref: "#/components/schemas/Label" } }
        checklist: { $ref: "#/components/schemas/ChecklistSummary" }
        commentCount: { type: integer }
        attachmentCount: { type: integer }
        position: { type: string }
        version: { type: integer, description: Optimistic concurrency token; send it back on update/move }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        completedAt: { type: string, format: date-time, nullable: true }
    CardInput:
      type: object
      required: [projectId, title]
      properties:
        projectId: { type: string, format: uuid }
        title: { type: string, minLength: 1, maxLength: 300 }
        description: { type: string }
        status: { $ref: "#/components/schemas/TaskStatus" }
        columnId: { type: string, format: uuid }
        priority: { $ref: "#/components/schemas/Priority" }
        dueDate: { type: string, format: date }
        assigneeIds: { type: array, items: { type: string, format: uuid } }
        labelIds: { type: array, items: { type: string, format: uuid }, maxItems: 10 }
    CardPatch:
      type: object
      required: [version]
      description: Absent fields are unchanged; dueDate null clears it.
      properties:
        version: { type: integer }
        title: { type: string }
        description: { type: string }
        priority: { $ref: "#/components/schemas/Priority" }
        dueDate: { type: string, format: date, nullable: true }
        assigneeIds: { type: array, items: { type: string, format: uuid } }
        labelIds: { type: array, items: { type: string, format: uuid }, maxItems: 10 }
    CardMove:
      type: object
      required: [version]
      description: |
        Target column (project board: neighbours must be in that column) or status only
        (cross-project board: neighbours may be any cards of that status; the card lands in its
        project's first column of the status). Omitted neighbours append to the end.
      properties:
        version: { type: integer }
        columnId: { type: string, format: uuid }
        status: { $ref: "#/components/schemas/TaskStatus" }
        afterId: { type: string, format: uuid, nullable: true, description: Card that will be directly above }
        beforeId: { type: string, format: uuid, nullable: true, description: Card that will be directly below }
    CardBoard:
      type: object
      required: [items, truncated]
      properties:
        items: { type: array, items: { $ref: "#/components/schemas/Card" } }
        truncated: { type: boolean, description: More cards match than were returned }
    Label:
      type: object
      required: [id, name, tone]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        tone: { $ref: "#/components/schemas/Tone" }
    LabelInput:
      type: object
      required: [name]
      properties:
        name: { type: string, minLength: 1, maxLength: 40 }
        tone: { $ref: "#/components/schemas/Tone" }
    LabelPatch:
      type: object
      properties:
        name: { type: string, minLength: 1, maxLength: 40 }
        tone: { $ref: "#/components/schemas/Tone" }
    ChecklistSummary:
      type: object
      required: [total, done]
      properties:
        total: { type: integer }
        done: { type: integer }
    ChecklistItem:
      type: object
      required: [id, text, done, position, completedAt]
      properties:
        id: { type: string, format: uuid }
        text: { type: string }
        done: { type: boolean }
        position: { type: string }
        completedAt: { type: string, format: date-time, nullable: true }
    ChecklistItemInput:
      type: object
      required: [text]
      properties:
        text: { type: string, minLength: 1, maxLength: 300 }
    ChecklistItemPatch:
      type: object
      description: Any subset; `move` reorders between neighbours.
      properties:
        text: { type: string, minLength: 1, maxLength: 300 }
        done: { type: boolean }
        move: { $ref: "#/components/schemas/Neighbours" }
    Neighbours:
      type: object
      properties:
        afterId: { type: string, format: uuid, nullable: true, description: Item that will be directly before }
        beforeId: { type: string, format: uuid, nullable: true, description: Item that will be directly after }
...
    CardPage:
      type: object
      required: [items, total, page, pageSize]
      properties:
        items: { type: array, items: { $ref: "#/components/schemas/Card" } }
        total: { type: integer }
        page: { type: integer }
        pageSize: { type: integer }
    StatusCounts:
      type: object
      required: [todo, in_progress, in_review, done]
      properties:
        todo: { type: integer }
        in_progress: { type: integer }
        in_review: { type: integer }
        done: { type: integer }
    BulkCardAction:
      type: object
      required: [ids, action]
      properties:
        ids: { type: array, items: { type: string, format: uuid }, maxItems: 500 }
        action: { type: string, enum: [move, priority, delete] }
        status: { $ref: "#/components/schemas/TaskStatus" }
        priority: { $ref: "#/components/schemas/Priority" }
    BulkResult:
      type: object
      required: [updated]
      properties:
        updated: { type: integer }
    DailyActivity:
      type: object
      required: [date, todo, in_progress, in_review, done]
      properties:
        date: { type: string, format: date }
        todo: { type: integer, description: Cards created }
        in_progress: { type: integer }
        in_review: { type: integer }
        done: { type: integer }
    ActivityItem:
      type: object
      required: [id, card, project, actor, from, to, at]
      properties:
        id: { type: integer, format: int64 }
        card:
          type: object
          required: [id, key, title]
          properties:
            id: { type: string, format: uuid }
            key: { type: string }
            title: { type: string }
        project: { $ref: "#/components/schemas/ProjectRef" }
        actor:
          allOf: [{ $ref: "#/components/schemas/PersonRef" }]
          nullable: true
        from:
          allOf: [{ $ref: "#/components/schemas/TaskStatus" }]
          nullable: true
        to: { $ref: "#/components/schemas/TaskStatus" }
        at: { type: string, format: date-time }
    Trend:
      type: object
      required: [value, previous, changePct]
      properties:
        value: { type: integer }
        previous: { type: integer }
        changePct: { type: number, description: "Percent change vs previous period (0 when previous is 0)" }
    DashboardStats:
      type: object
      required: [active, total, inReview, overdue, completedThisWeek, createdThisWeek, statusCounts, daily, activity]
      properties:
        active: { type: integer }
        total: { type: integer }
        inReview: { type: integer }
        overdue: { type: integer }
        completedThisWeek: { $ref: "#/components/schemas/Trend" }
        createdThisWeek: { $ref: "#/components/schemas/Trend" }
        statusCounts: { $ref: "#/components/schemas/StatusCounts" }
        daily: { type: array, items: { $ref: "#/components/schemas/DailyActivity" } }
        activity: { type: array, items: { $ref: "#/components/schemas/ActivityItem" } }
security:
  - cookieAuth: []